Simplify parser grammar, and remove comparison atom names.
The parser is inherently ambiguous, if you separate boolean, string and arithmetic expressions.
We merge the three types and create operations that work on any of them. This is a superset of
the valid expressions and validation is then made as far as possible by a separate type scan of
the parse tree.  Full validation is not possible until run-time when the types of the identifiers
is known.

This commit also removes the atoms 'eq', 'neq', 'lteq', &c. and replaces them with direct atoms of
the operator, viz. '=', '<>', '<=', &c.

[v1.0.x]